implement new packets, handlers and database queries
- Implemented many req, res, ntc and handlers for Epitaph content.
- Renamed existing BonusDungeonManager to DungeonManager.
- Implemented database migration from 24 -> 25.
- Implemented a new packet handler which provides a queue which can be
  used to send packets after the response is returned.
- Commented out RPC ping print message.
- Added new NPC behavior to NpcGetExtendedFacilityHandler.

+++ Arrowgene.Ddon.Shared/Model/Quest/QuestEnemyGroup.cs
@@ -4,7 +4,7 @@ namespace Arrowgene.Ddon.Shared.Model.Quest
 {
     public class QuestEnemyGroup
     {
-        public uint SubGroupId { get; set; }
+        public byte SubGroupId { get; set; }
         public StageId StageId { get; set; }
         public uint StartingIndex { get; set; }
         public List<InstancedEnemy> Enemies { get; set; }
